feat: replace opentracing with otel
