+++ server/v1/module.go
@@ -2,7 +2,7 @@ package v1
 
 import (
 	"github.com/alexfalkowski/migrieren/migrate"
-	"github.com/alexfalkowski/migrieren/migrate/trace/opentracing"
+	"github.com/alexfalkowski/migrieren/migrate/otel"
 	"github.com/alexfalkowski/migrieren/server/v1/transport/grpc"
 	"go.uber.org/fx"
 )
@@ -11,7 +11,7 @@ var (
 	// Module for fx.
 	Module = fx.Options(
 		fx.Provide(migrate.NewMigrator),
-		fx.Provide(opentracing.NewTracer),
+		fx.Provide(otel.NewTracer),
 		fx.Provide(grpc.NewServer),
 		fx.Invoke(grpc.Register),
 	)
