+++ test/.config/misconfigured.client.yml
@@ -4,10 +4,9 @@ client:
     timeout: 1s
     database: invalid_source
     version: 1
-trace:
-  opentracing:
-    kind: jaeger
-    host: localhost:6831
+otel:
+  kind: jaeger
+  host: localhost:6831
 transport:
   port: 8080
   grpc:
