+++ test/.config/valid.client.yml
@@ -4,10 +4,9 @@ client:
     timeout: 1s
     database: postgres
     version: 1
-trace:
-  opentracing:
-    kind: jaeger
-    host: localhost:6831
+otel:
+  kind: jaeger
+  host: localhost:6831
 transport:
   port: 8080
   grpc:
